# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API fashion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/pieces": {
            "get": {
                "summary": "Show pieces",
                "operationId": "d1b8238e59eac61a43053323ca31e47b",
                "responses": {
                    "200": {
                        "description": "show all pieces.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                }
            },
            "post": {
                "summary": "Create piece",
                "operationId": "812fd41a59d234853c03b2a92374147f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the piece",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type of the piece",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color of the piece",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the piece",
                                        "type": "number"
                                    },
                                    "status": {
                                        "description": "status of the piece",
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "description": "Category ID of the piece",
                                        "type": "number"
                                    },
                                    "file": {
                                        "description": "Image file of the piece",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Piece successfully created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pieces/{id}": {
            "post": {
                "summary": "Update piece",
                "operationId": "b79886044a7debc74e9dcfd0a43f220a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "type",
                                    "color",
                                    "price",
                                    "status",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the piece",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type of the piece",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Color of the piece",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the piece",
                                        "type": "number"
                                    },
                                    "status": {
                                        "description": "Status of the piece",
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "description": "Category ID of the piece",
                                        "type": "number"
                                    },
                                    "file": {
                                        "description": "Image file of the piece. Only jpg, png, jpeg, gif, svg files are allowed.",
                                        "type": "string",
                                        "format": "binary",
                                        "pattern": "^.+\\.(jpg|png|jpeg|gif|svg)$"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Piece successfully updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete piece",
                "operationId": "57063793f2e07d3532ea07d49c9edf85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "piece successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    }
}